Got testing off of "single allowed" vs "assigned" working. 2x single_line test improvement (3s->1.5s)

+++ puzzle/solution_permuter/allowed_value_solution_permuter.h
@@ -59,7 +59,7 @@ class AllowedValueGrid {
 
  private:
   friend class AllowedValueAdvancer;
-  int CheckAllowed(SolutionFilter filter, Box box) const;
+  uint16_t CheckAllowed(SolutionFilter filter, Box box) const;
 
   // Returns true if there are still potential solutions.
   bool Prepare();
